wip decoupling asset upload action from main designer route
