+++ apps/designer/app/shared/db/misc.server.ts
@@ -1,4 +1,3 @@
-export { uploadAssets } from "@webstudio-is/asset-uploader";
 import * as db from "./";
 /**
  * Conceptually publishing is cloning all data that affects user site
